MockEmbedderApiForKeyboard
