+++ shell/platform/windows/testing/test_keyboard.cc
@@ -2,8 +2,11 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
+#include "flutter/shell/platform/common/json_message_codec.h"
+#include "flutter/shell/platform/embedder/test_utils/proc_table_replacement.h"
 #include "flutter/shell/platform/windows/testing/test_keyboard.h"
 
+#include <rapidjson/document.h>
 #include <windowsx.h>
 
 namespace flutter {
@@ -26,6 +29,23 @@ std::string _print_character(const char* s) {
   }
   return std::string("\"") + s + "\"";
 }
+
+std::unique_ptr<std::vector<uint8_t>> _keyHandlingResponse(bool handled) {
+  rapidjson::Document document;
+  auto& allocator = document.GetAllocator();
+  document.SetObject();
+  document.AddMember("handled", handled, allocator);
+  return flutter::JsonMessageCodec::GetInstance().EncodeMessage(document);
+}
+
+// A struct to use as a FlutterPlatformMessageResponseHandle so it can keep the
+// callbacks and user data passed to the engine's
+// PlatformMessageCreateResponseHandle for use in the SendPlatformMessage
+// overridden function.
+struct TestResponseHandle {
+  FlutterDesktopBinaryReply callback;
+  void* user_data;
+};
 }  // namespace
 
 #define _RETURN_IF_NOT_EQUALS(val1, val2)                                     \
@@ -63,6 +83,87 @@ LPARAM CreateKeyEventLparam(USHORT scancode,
   return ((LPARAM(transition_state) << 31) | (LPARAM(was_down) << 30) |
           (LPARAM(context_code) << 29) | (LPARAM(extended ? 0x1 : 0x0) << 24) |
           (LPARAM(scancode) << 16) | LPARAM(repeat_count));
+}
+
+void MockEmbedderApiForKeyboard(EngineModifier& modifier,
+                                MockKeyEventChannelHandler channel_handler,
+                                MockKeyEventEmbedderHandler embedder_handler) {
+  // This mock handles channel messages.
+  modifier.embedder_api().SendPlatformMessage = MOCK_ENGINE_PROC(
+      SendPlatformMessage,
+      [channel_handler](FLUTTER_API_SYMBOL(FlutterEngine) engine,
+                        const FlutterPlatformMessage* message) {
+        if (std::string(message->channel) == std::string("flutter/settings")) {
+          return kSuccess;
+        }
+        if (std::string(message->channel) == std::string("flutter/keyevent")) {
+          bool result = channel_handler();
+          auto response = _keyHandlingResponse(result);
+          const TestResponseHandle* response_handle =
+              reinterpret_cast<const TestResponseHandle*>(
+                  message->response_handle);
+          if (response_handle->callback != nullptr) {
+            response_handle->callback(response->data(), response->size(),
+                                      response_handle->user_data);
+          }
+          return kSuccess;
+        }
+        return kSuccess;
+      });
+
+  // This mock handles key events sent through the embedder API,
+  // and records it in `key_calls`.
+  modifier.embedder_api().SendKeyEvent = MOCK_ENGINE_PROC(
+      SendKeyEvent,
+      [embedder_handler](FLUTTER_API_SYMBOL(FlutterEngine) engine, const FlutterKeyEvent* event,
+         FlutterKeyEventCallback callback, void* user_data) {
+        bool result = embedder_handler(event);
+        if (callback != nullptr) {
+          callback(result, user_data);
+        }
+        return kSuccess;
+      });
+
+  // The following mocks enable channel mocking.
+  modifier.embedder_api().PlatformMessageCreateResponseHandle =
+      [](auto engine, auto data_callback, auto user_data, auto response_out) {
+        TestResponseHandle* response_handle = new TestResponseHandle();
+        response_handle->user_data = user_data;
+        response_handle->callback = data_callback;
+        *response_out = reinterpret_cast<FlutterPlatformMessageResponseHandle*>(
+            response_handle);
+        return kSuccess;
+      };
+
+  modifier.embedder_api().PlatformMessageReleaseResponseHandle =
+      [](FLUTTER_API_SYMBOL(FlutterEngine) engine,
+         FlutterPlatformMessageResponseHandle* response) {
+        const TestResponseHandle* response_handle =
+            reinterpret_cast<const TestResponseHandle*>(response);
+        delete response_handle;
+        return kSuccess;
+      };
+
+  // The following mocks allows RunWithEntrypoint to be run, which creates a
+  // non-empty FlutterEngine and enables SendKeyEvent.
+
+  modifier.embedder_api().Run =
+      [](size_t version, const FlutterRendererConfig* config,
+         const FlutterProjectArgs* args, void* user_data,
+         FLUTTER_API_SYMBOL(FlutterEngine) * engine_out) {
+        *engine_out = reinterpret_cast<FLUTTER_API_SYMBOL(FlutterEngine)>(1);
+
+        return kSuccess;
+      };
+  modifier.embedder_api().UpdateLocales =
+      [](auto engine, const FlutterLocale** locales, size_t locales_count) {
+        return kSuccess;
+      };
+  modifier.embedder_api().SendWindowMetricsEvent =
+      [](auto engine, const FlutterWindowMetricsEvent* event) {
+        return kSuccess;
+      };
+  modifier.embedder_api().Shutdown = [](auto engine) { return kSuccess; };
 }
 
 }  // namespace testing
